Almost Done

+++ src/model/product/Digital.java
@@ -3,7 +3,7 @@ package model.product;
 import model.interfaces.Shippable;
 
 public class Digital extends Product {
-    public Digital(String name, double price, int quantity) {
+    public Digital(String name, double price, Long quantity) {
         super(name, price, quantity);
     }
